# test: added small python tests to see performance