--- a/auth/src/main/java/org/example/authserver/service/RelationsService.java
+++ b/auth/src/main/java/org/example/authserver/service/RelationsService.java
@@ -1,6 +1,7 @@
 package org.example.authserver.service;
 
 import io.micrometer.core.annotation.Timed;
+import io.micrometer.core.instrument.MeterRegistry;
 import lombok.extern.slf4j.Slf4j;
 import org.example.authserver.service.model.RequestCache;
 import org.example.authserver.service.zanzibar.Zanzibar;
@@ -15,10 +16,12 @@ public class RelationsService {
 
     private final Zanzibar zanzibar;
     private final UserRelationsCacheService userRelationsCacheService;
+    private final MeterRegistry meterRegistry;
 
-    public RelationsService(Zanzibar zanzibar, UserRelationsCacheService userRelationsCacheService) {
+    public RelationsService(Zanzibar zanzibar, UserRelationsCacheService userRelationsCacheService, MeterRegistry meterRegistry) {
         this.zanzibar = zanzibar;
         this.userRelationsCacheService = userRelationsCacheService;
+        this.meterRegistry = meterRegistry;
     }
 
     @Timed(value = "relation.get", percentiles = {0.99, 0.95, 0.75})
@@ -29,11 +32,6 @@ public class RelationsService {
             return cachedRelations.get();
         }
 
-        return getZanzibarRelations(namespace, object, principal, requestCache);
-    }
-
-    @Timed(value = "relation.zanzibar", percentiles = {0.99, 0.95, 0.75})
-    public Set<String> getZanzibarRelations(String namespace, String object, String principal, RequestCache requestCache) {
-        return zanzibar.getRelations(namespace, object, principal, requestCache);
+        return meterRegistry.timer("relation.zanzibar").record(() -> zanzibar.getRelations(namespace, object, principal, requestCache));
     }
 }