--- a/auth/src/main/java/org/example/authserver/service/UserRelationCacheBuilder.java
+++ b/auth/src/main/java/org/example/authserver/service/UserRelationCacheBuilder.java
@@ -1,6 +1,7 @@
 package org.example.authserver.service;
 
 import com.google.common.base.Stopwatch;
+import io.micrometer.core.annotation.Timed;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.example.authserver.config.UserRelationsConfig;
@@ -139,6 +140,7 @@ public class UserRelationCacheBuilder {
         buildUserRelations(user, namespaces, objects);
     }
 
+    @Timed(value = "cache.buildUserRelations", percentiles = {0.99, 0.95, 0.75})
     public void buildUserRelations(String user, Set<String> namespaces, Set<String> objects) {
         Optional<UserRelationEntity> entityOptional = createUserRelations(user, namespaces, objects);
         if (entityOptional.isEmpty()) {