--- a/auth/src/main/java/org/example/authserver/service/zanzibar/ZanzibarImpl.java
+++ b/auth/src/main/java/org/example/authserver/service/zanzibar/ZanzibarImpl.java
@@ -44,7 +44,7 @@ public class ZanzibarImpl implements Zanzibar {
     }
 
     @Override
-    @Timed(value = "relation.zanzibar", percentiles = {0.99, 0.95, 0.75})
+    @Timed(value = "getRelation", percentiles = {0.99, 0.95, 0.75})
     public Set<String> getRelations(String namespace, String object, String principal, RequestCache requestCache) {
         Set<ExpandedAcl> relations = expandMultiple(Set.of(Tuples.of(namespace, object)), principal, requestCache);
         Set<Tuple2<String, String>> lookups = lookup(relations, namespace, object, principal);