--- a/auth/src/main/resources/application-dev.properties
+++ b/auth/src/main/resources/application-dev.properties
@@ -8,9 +8,9 @@ app.user-relations-cache.scheduled-period-time-unit=MINUTES
 
 
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/authz
-spring.datasource.username=postgres
-spring.datasource.password=secret
+spring.datasource.url=jdbc:postgresql://localhost:5432/lendistry
+spring.datasource.username=sa
+spring.datasource.password=password
 spring.datasource.hikari.maximumPoolSize=1
 
 spring.jpa.database=postgresql