--- a/perftest/src/main/python/debug_contoller_test.py
+++ b/perftest/src/main/python/debug_contoller_test.py
@@ -3,7 +3,7 @@ import sys
 from locust import HttpUser, task
 
 # Run as:
-# locust --headless --users 3 --spawn-rate 1 --host http://localhost:8183/debug  -f .\debug_contoller_test.py
+# locust --headless --users 3 --spawn-rate 1 --host http://localhost:8183  -f .\debug_contoller_test.py
 class AuthTest(HttpUser):
     service_url = ""
     auth = ""
@@ -32,4 +32,4 @@ class AuthTest(HttpUser):
         }
 
         params = {"namespace": namespace, "object": object, "relation": relation, "principal": principal}
-        return self.client.get("/test", headers=headers, params=params)
+        return self.client.get("/debug/test", headers=headers, params=params)