test: added small python tests to see performance

If cache in place then : 

- fetching relations takes up to 6ms, ~200 req/s
- add acl takes around 8ms, ~200 req/s

**Debug check relations for principal** - `debug_contoller_test.py`
```
[2021-10-29 18:03:08,874] DELL-Inspiron15/INFO/locust.main: Cleaning up runner...
 Name                                                                              # reqs      # fails  |     Avg     Min     Max  Median  |   req/s failures/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 GET /debug/test?namespace=test&object=coarse-access&relation=TB&principal=0003d627-3a09-4891-84ff-8c624dcefa76     467     0(0.00%)  |       7       3      68       7  |  202.91    0.00
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregated                                                                           467     0(0.00%)  |       7       3      68       7  |  202.91    0.00

```

```
2021-10-29 18:06:56.703  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 3ms
2021-10-29 18:06:56.705  INFO 6568 --- [  XNIO-1 task-2] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 3ms
2021-10-29 18:06:56.707  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 3ms
2021-10-29 18:06:56.710  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 3ms
2021-10-29 18:06:56.712  INFO 6568 --- [  XNIO-1 task-2] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.715  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.717  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.720  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 5ms
2021-10-29 18:06:56.724  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.728  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 6ms
2021-10-29 18:06:56.732  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 6ms
2021-10-29 18:06:56.735  INFO 6568 --- [  XNIO-1 task-2] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 5ms
2021-10-29 18:06:56.738  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 6ms
2021-10-29 18:06:56.741  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 5ms
2021-10-29 18:06:56.743  INFO 6568 --- [  XNIO-1 task-2] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 5ms
2021-10-29 18:06:56.746  INFO 6568 --- [  XNIO-1 task-1] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.748  INFO 6568 --- [  XNIO-1 task-3] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 4ms
2021-10-29 18:06:56.752  INFO 6568 --- [  XNIO-1 task-2] o.e.a.controller.DebugController         : get relations: test:coarse-access @ 89a13a86-d5c2-4701-9ac7-2d8402e1f1d3, 5ms
```

**Add acl** - `add_acl.py`
```
2021-10-29T15:24:40Z <Greenlet at 0x24c14f0ea68: run_user(<add_acl.AuthTest object at 0x0000024C14F54A88>)> failed with KeyboardInterrupt

[2021-10-29 18:24:40,316] DELL-Inspiron15/INFO/locust.main: Running teardowns...
[2021-10-29 18:24:40,316] DELL-Inspiron15/INFO/locust.main: Shutting down (exit code 0), bye.
[2021-10-29 18:24:40,316] DELL-Inspiron15/INFO/locust.main: Cleaning up runner...
 Name                                                                              # reqs      # fails  |     Avg     Min     Max  Median  |   req/s failures/s
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 POST /acl/create                                                                     926     0(0.00%)  |      11       4     188      11  |  203.98    0.00
----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Aggregated                                                                           926     0(0.00%)  |      11       4     188      11  |  203.98    0.00

Response time percentiles (approximated)
 Type     Name                                                                                  50%    66%    75%    80%    90%    95%    98%    99%  99.9% 99.99%   100% # reqs
--------|--------------------------------------------------------------------------------|---------|------|------|------|------|------|------|------|------|------|------|------|
 POST     /acl/create                                                                            11     12     12     13     14     15     17     20    190    190    190    926
--------|--------------------------------------------------------------------------------|---------|------|------|------|------|------|------|------|------|------|------|------|
 None     Aggregated                                                                             11     12     12     13     14     15     17     20    190    190    190    926

PS C:\U\projects\p\envoy-authz\perftest\src\main\python>

```
```
2021-10-29 18:24:40.243  INFO 2500 --- [  XNIO-1 task-3] o.e.authserver.controller.AclController  : Created ACL: Acl(id=80f126df-4d7c-495b-9e86-0626c46cffc8, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080239, updated=1635521080239), time 4ms
2021-10-29 18:24:40.245  WARN 2500 --- [pool-1-thread-2] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.245  INFO 2500 --- [  XNIO-1 task-1] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.245  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Created ACL: Acl(id=1078b63f-60b0-4652-9211-795a0bf110a2, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080241, updated=1635521080241), time 3ms
2021-10-29 18:24:40.245  INFO 2500 --- [  XNIO-1 task-3] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=7465f0de-a09b-4c4b-8688-fa3bff01fd59, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080245, updated=1635521080245)
2021-10-29 18:24:40.247  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=5dd793ff-cb49-4d38-bda3-999469478ca1, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080247, updated=1635521080247)
2021-10-29 18:24:40.248  INFO 2500 --- [  XNIO-1 task-3] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.248  WARN 2500 --- [pool-1-thread-2] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.248  INFO 2500 --- [  XNIO-1 task-3] o.e.authserver.controller.AclController  : Created ACL: Acl(id=7465f0de-a09b-4c4b-8688-fa3bff01fd59, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080245, updated=1635521080245), time 3ms
2021-10-29 18:24:40.251  INFO 2500 --- [  XNIO-1 task-3] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=cb690ed2-d4bf-4836-81a2-f17579f0fec0, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080250, updated=1635521080250)
2021-10-29 18:24:40.252  INFO 2500 --- [  XNIO-1 task-1] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.252  WARN 2500 --- [pool-1-thread-2] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.252  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Created ACL: Acl(id=5dd793ff-cb49-4d38-bda3-999469478ca1, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080247, updated=1635521080247), time 5ms
2021-10-29 18:24:40.255  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=e5f8125f-c2d7-4fa1-bc1f-10cac05cc12e, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080254, updated=1635521080254)
2021-10-29 18:24:40.257  INFO 2500 --- [  XNIO-1 task-2] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=c32b6110-45b1-4542-bf60-a41593b9e9c8, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080257, updated=1635521080257)
2021-10-29 18:24:40.259  INFO 2500 --- [  XNIO-1 task-3] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.259  INFO 2500 --- [  XNIO-1 task-3] o.e.authserver.controller.AclController  : Created ACL: Acl(id=cb690ed2-d4bf-4836-81a2-f17579f0fec0, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080250, updated=1635521080250), time 8ms
2021-10-29 18:24:40.260  WARN 2500 --- [pool-1-thread-1] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.260  INFO 2500 --- [  XNIO-1 task-1] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.260  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Created ACL: Acl(id=e5f8125f-c2d7-4fa1-bc1f-10cac05cc12e, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080254, updated=1635521080254), time 5ms
2021-10-29 18:24:40.261  INFO 2500 --- [  XNIO-1 task-2] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.261  WARN 2500 --- [pool-1-thread-1] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.261  INFO 2500 --- [  XNIO-1 task-2] o.e.authserver.controller.AclController  : Created ACL: Acl(id=c32b6110-45b1-4542-bf60-a41593b9e9c8, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080257, updated=1635521080257), time 3ms
2021-10-29 18:24:40.262  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=c2451591-ba1b-4563-9e46-a18d154311ea, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080261, updated=1635521080261)
2021-10-29 18:24:40.265  INFO 2500 --- [  XNIO-1 task-2] o.e.authserver.controller.AclController  : Creating ACL: Acl(id=f46322bc-9d0b-416c-bcc5-b75208570982, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080265, updated=1635521080265)
2021-10-29 18:24:40.266  INFO 2500 --- [  XNIO-1 task-1] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.266  WARN 2500 --- [pool-1-thread-1] o.e.a.service.UserRelationCacheBuilder   : Building for user null is already in progress. Scheduled update for later.
2021-10-29 18:24:40.266  INFO 2500 --- [  XNIO-1 task-1] o.e.authserver.controller.AclController  : Created ACL: Acl(id=c2451591-ba1b-4563-9e46-a18d154311ea, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080261, updated=1635521080261), time 4ms
2021-10-29 18:24:40.275  INFO 2500 --- [  XNIO-1 task-2] o.e.a.service.UserRelationCacheBuilder   : Scheduled updated for user null.
2021-10-29 18:24:40.275  INFO 2500 --- [  XNIO-1 task-2] o.e.authserver.controller.AclController  : Created ACL: Acl(id=f46322bc-9d0b-416c-bcc5-b75208570982, namespace=test, object=TB, relation=coarse-access, user=null, usersetNamespace=null, usersetObject=null, usersetRelation=null, created=1635521080265, updated=1635521080265), time 9ms

```