@yuriyz-lendistry 2021-11-01
Added metrics which should be for informative. 
1. `relation.get` - relations are get
2. `relation.zanzibar` - relation invoked zanzibar for building
3. `relation.cache` - creating relation db cache

Sample
```
relation_get_seconds_count{class="org.example.authserver.service.RelationsService",exception="none",method="getRelations",} 3031.0
relation_zanzibar_seconds_count 898.0
relation_cache_build_count{class="org.example.authserver.service.UserRelationCacheBuilder",exception="none",method="getRelations",} 1
```

Means we got totally 3031 calls from which 898 were not cached and invoked zanzibar.